+++ NEWS.md
@@ -10,6 +10,11 @@
     * Default adduct can be specified in `newProject()` dialog
 * Small compatibility fix for feature finding with OpenMS 2.5 (reported by Thanh Wang)
 * RAMClustR is now supported from CRAN and no need to install github package anymore
+* pubchemlite identifiers are now URL linked in HTML reports
+* related CIDs are now reported for PubChemLite results.
+* MetFrag compound generation: removed `addTrivialNames` option as it never worked very well.
+* `reportHTML()`: only components with reported feature groups are now reported.
+
 
 ## February 2020